requirements.txt
django>=1.8.18
channels>=1.1.3
djangorestframework>=3.4.1
